+++ code/src/sixsq/nuvla/server/resources/evidence_record.clj
@@ -36,7 +36,7 @@ justifying whether a particular check within the process has passed or failed.
   (let [valid-prefixes   (sn/all-prefixes)
         resource-payload (dissoc resource
                                  :acl :id :resource-type :name :description
-                                 :created :updated :properties :operations
+                                 :created :updated :tags :operations
                                  :class :plan-id :start-time :end-time :passed)
         validator        (partial key-utils/valid-attribute-name? valid-prefixes)]
     (if (key-utils/valid-attributes? validator resource-payload)