+++ code/src/sixsq/nuvla/server/resources/nuvlabox_peripheral.clj
@@ -1,0 +1,112 @@
+(ns sixsq.nuvla.server.resources.nuvlabox-peripheral
+  "
+The nuvlabox-peripheral resource represents a peripheral attached to a
+nuvlabox.
+"
+  (:require
+    [sixsq.nuvla.auth.acl-resource :as a]
+    [sixsq.nuvla.server.resources.common.crud :as crud]
+    [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
+    [sixsq.nuvla.server.resources.common.utils :as u]
+    [sixsq.nuvla.server.resources.resource-metadata :as md]
+    [sixsq.nuvla.server.resources.spec.nuvlabox-peripheral :as nb-peripheral]
+    [sixsq.nuvla.server.util.metadata :as gen-md]
+    [sixsq.nuvla.server.util.response :as r]))
+
+
+(def ^:const resource-type (u/ns->type *ns*))
+
+
+(def ^:const collection-type (u/ns->collection-type *ns*))
+
+
+(def collection-acl {:add   ["group/nuvla-user"]
+                     :query ["group/nuvla-user"]})
+
+
+;;
+;; multimethods for validation
+;;
+
+(defmulti validate-subtype
+          "Validates the given nuvlabox-peripheral resource against a specific
+           version of the schema."
+          :version)
+
+
+(defmethod validate-subtype :default
+  [{:keys [version] :as resource}]
+  (if version
+    (throw (r/ex-bad-request (str "unsupported nuvlabox-peripheral version: " version)))
+    (throw (r/ex-bad-request "missing nuvlabox-peripheral version"))))
+
+
+(defmethod crud/validate resource-type
+  [resource]
+  (validate-subtype resource))
+
+;;
+;; acl
+;;
+
+(defmethod crud/add-acl resource-type
+  [resource request]
+  (-> (a/add-acl resource request)
+      (update-in [:acl :view-data] (comp vec conj) "group/nuvla-user")))
+
+
+;;
+;; CRUD operations
+;;
+
+(def add-impl (std-crud/add-fn resource-type collection-acl resource-type))
+
+
+(defmethod crud/add resource-type
+  [request]
+  (add-impl request))
+
+
+(def edit-impl (std-crud/edit-fn resource-type))
+
+
+(defmethod crud/edit resource-type
+  [request]
+  (edit-impl request))
+
+
+(def retrieve-impl (std-crud/retrieve-fn resource-type))
+
+
+(defmethod crud/retrieve resource-type
+  [request]
+  (retrieve-impl request))
+
+
+(def delete-impl (std-crud/delete-fn resource-type))
+
+
+(defmethod crud/delete resource-type
+  [request]
+  (delete-impl request))
+
+
+(def query-impl (std-crud/query-fn resource-type collection-acl collection-type))
+
+
+(defmethod crud/query resource-type
+  [request]
+  (query-impl request))
+
+
+;;
+;; initialization
+;;
+
+(def resource-metadata (gen-md/generate-metadata ::ns ::nb-peripheral/schema))
+
+
+(defn initialize
+  []
+  (std-crud/initialize resource-type ::nb-peripheral/schema)
+  (md/register resource-metadata))