+++ code/test/sixsq/nuvla/server/resources/nuvlabox_0_lifecycle_test.clj
@@ -51,10 +51,8 @@
                                         :view-acl ["user/jane"]
                                         :manage   ["user/jane"]}
 
-                     ;; This doesn't need to be specified as it will default to the
-                     ;; latest version (which is currently 0). If new versions are added,
-                     ;; the following line must be uncommented.
-                     ;:version          0
+                     ;; This is not the default version, so it must appear explicitly.
+                     :version          0
 
                      :owner            nuvlabox-owner
                      :organization     "ACME"