+++ code/test/sixsq/nuvla/server/resources/user_template_lifecycle_test.clj
@@ -10,6 +10,8 @@
     [sixsq.nuvla.server.resources.user-template :as t]
     [sixsq.nuvla.server.resources.user-template-email-invitation :as email-invitation]
     [sixsq.nuvla.server.resources.user-template-email-password :as email-password]
+    [sixsq.nuvla.server.resources.user-template-github :as github]
+    [sixsq.nuvla.server.resources.user-template-minimum :as minimum]
     [sixsq.nuvla.server.resources.user-template-username-password :as username-password]
     [sixsq.nuvla.server.util.metadata-test-utils :as mdtu]))
 
@@ -22,6 +24,8 @@
 
 (deftest check-metadata
   (mdtu/check-metadata-exists t/resource-type
+                              (str t/resource-type "-" github/resource-url)
+                              (str t/resource-type "-" github/resource-url "-create")
                               (str t/resource-type "-" email-password/resource-url)
                               (str t/resource-type "-" email-password/resource-url "-create")
                               (str t/resource-type "-" username-password/resource-url)
@@ -55,12 +59,14 @@
         ids     (set (map :id entries))
         types   (set (map :method entries))]
 
-    (is (= #{(str t/resource-type "/" email-password/registration-method)
+    (is (= #{(str t/resource-type "/" minimum/registration-method)
+             (str t/resource-type "/" email-password/registration-method)
              (str t/resource-type "/" username-password/registration-method)
              (str t/resource-type "/" email-invitation/registration-method)}
            ids))
 
-    (is (= #{email-password/registration-method
+    (is (= #{minimum/registration-method
+             email-password/registration-method
              username-password/registration-method
              email-invitation/registration-method}
            types))